+++ lib/cors_validator.dart
@@ -1,0 +1,6 @@
+library cors_validator;
+
+export 'src/factories/cors_validator_factory.dart';
+export 'src/domain/enums/package_requisition_enum.dart';
+export 'src/domain/entities/cors_validation_result_entity.dart';
+export 'src/domain/usecases/validate_cors_usecase.dart';