+++ lib/src/domain/entities/cors_validation_result_entity.dart
@@ -1,0 +1,13 @@
+class CorsValidationResultEntity {
+  final bool isValid;
+  final int statusCode;
+  final String? body;
+  final String? errorMessage;
+
+  CorsValidationResultEntity({
+    required this.isValid,
+    required this.statusCode,
+    this.body,
+    this.errorMessage,
+  });
+}