+++ lib/src/domain/enums/package_requisition_enum.dart
@@ -1,0 +1,1 @@
+enum PackageRequisitionEnum { dio, http, unsupported }