+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release of `cors_validator`.
+* Added support for validating CORS using Dio and http clients.
+* Provided a factory (`CorsValidatorFactory`) to create instances of `CorsValidator`.
+* Included examples for using `cors_validator` in Dart and Flutter applications.
+* Added documentation and usage examples in the `README.md`.